# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recipes API",
        "description": "API para gerenciamento de receitas culinárias",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/public/recipes": {
            "get": {
                "tags": [
                    "Public Recipes"
                ],
                "summary": "Get all public recipes with filters",
                "description": "Display a listing of all public recipes with filters.",
                "operationId": "5ed8d1ce60af07375b908799705c0ba7",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "servings_operator",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "exact",
                                "above",
                                "below"
                            ]
                        },
                        "example": "exact"
                    },
                    {
                        "name": "servings_value",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 4
                    },
                    {
                        "name": "prep_time_operator",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "exact",
                                "above",
                                "below"
                            ]
                        },
                        "example": "below"
                    },
                    {
                        "name": "prep_time_value",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 30
                    },
                    {
                        "name": "rating_operator",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "exact",
                                "above",
                                "below"
                            ]
                        },
                        "example": "above"
                    },
                    {
                        "name": "rating_value",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        },
                        "example": 4
                    },
                    {
                        "name": "comments_operator",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "exact",
                                "above",
                                "below"
                            ]
                        },
                        "example": "above"
                    },
                    {
                        "name": "comments_value",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 5
                    },
                    {
                        "name": "my_recipes",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": false
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "bolo"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipes retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Recipe"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 150
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/public/recipes/{id}": {
            "get": {
                "tags": [
                    "Public Recipes"
                ],
                "summary": "Get a specific public recipe",
                "description": "Display the specified public recipe.",
                "operationId": "e8e183049c00a85256f91ed7becd095c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipe retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Recipe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recipe not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Recipe": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "category": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Sobremesas"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Bolo de Chocolate",
                        "nullable": true
                    },
                    "prep_time_minutes": {
                        "type": "integer",
                        "example": 30,
                        "nullable": true
                    },
                    "servings": {
                        "type": "integer",
                        "example": 8,
                        "nullable": true
                    },
                    "image": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/image.jpg",
                        "nullable": true
                    },
                    "instructions": {
                        "type": "string",
                        "example": "Misture todos os ingredientes..."
                    },
                    "ingredients": {
                        "type": "string",
                        "example": "2 xícaras de farinha, 3 ovos...",
                        "nullable": true
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "comment": {
                                    "type": "string",
                                    "example": "Delicioso!"
                                },
                                "user": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        }
                                    },
                                    "type": "object"
                                },
                                "created_at": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            },
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "comments_count": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "ratings_count": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "average_rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}